check payment of enough akt
